feat: refactor promotion discount plugin
