+++ apps/reaction/plugins.json
@@ -25,8 +25,6 @@
   "payments": "@reactioncommerce/api-plugin-payments",
   "paymentsStripeSCA": "@reactioncommerce/api-plugin-payments-stripe-sca",
   "paymentsExample": "@reactioncommerce/api-plugin-payments-example",
-  "discounts": "@reactioncommerce/api-plugin-discounts",
-  "discountCodes": "@reactioncommerce/api-plugin-discounts-codes",
   "surcharges": "@reactioncommerce/api-plugin-surcharges",
   "shipments": "@reactioncommerce/api-plugin-shipments",
   "shipmentsFlatRate": "@reactioncommerce/api-plugin-shipments-flat-rate",
