+++ packages/api-plugin-promotions-discounts/src/actions/discountAction.test.js
@@ -1,7 +1,7 @@
 import applyItemDiscountToCart from "../utils/discountTypes/item/applyItemDiscountToCart.js";
 import applyOrderDiscountToCart from "../utils/discountTypes/order/applyOrderDiscountToCart.js";
 import applyShippingDiscountToCart from "../utils/discountTypes/shipping/applyShippingDiscountToCart.js";
-import discountAction, { discountActionHandler, discountActionParameters } from "./discountAction.js";
+import discountAction, { discountActionCleanup, discountActionHandler, discountActionParameters } from "./discountAction.js";
 
 jest.mock("../utils/discountTypes/item/applyItemDiscountToCart.js", () => jest.fn());
 jest.mock("../utils/discountTypes/order/applyOrderDiscountToCart.js", () => jest.fn());
@@ -54,4 +54,50 @@ test("should call discount shipping function when discountType parameters is shi
   };
   discountAction.handler(context, cart, params);
   expect(applyShippingDiscountToCart).toHaveBeenCalledWith(context, params, cart);
+});
+
+describe("cleanup", () => {
+  test("should reset the cart discount state", () => {
+    const cart = {
+      discounts: [{ _id: "discount1" }],
+      discount: 10,
+      items: [
+        {
+          _id: "item1",
+          discounts: [{ _id: "discount1" }],
+          price: {
+            amount: 12
+          },
+          quantity: 1,
+          subtotal: {
+            amount: 10,
+            currencyCode: "USD",
+            discount: 2,
+            undiscountedAmount: 12
+          }
+        }
+      ]
+    };
+
+    const updatedCart = discountActionCleanup({}, cart);
+
+    expect(updatedCart).toEqual({
+      discounts: [],
+      discount: 0,
+      items: [
+        {
+          _id: "item1",
+          discounts: [],
+          price: {
+            amount: 12
+          },
+          quantity: 1,
+          subtotal: {
+            amount: 12,
+            currencyCode: "USD"
+          }
+        }
+      ]
+    });
+  });
 });
