+++ packages/api-plugin-promotions-discounts/src/discountTypes/item/applyItemDiscountToCart.test.js
@@ -185,3 +185,83 @@ test("should return cart with applied discount when parameters include rule", as
     discountedItems: [item]
   });
 });
+
+describe("recalculateCartItemSubtotal", () => {
+  test("should recalculate the item subtotal with discountType is item", () => {
+    const item = {
+      _id: "item1",
+      price: {
+        amount: 12
+      },
+      quantity: 1,
+      subtotal: {
+        amount: 10,
+        currencyCode: "USD",
+        discount: 2,
+        undiscountedAmount: 12
+      },
+      discounts: []
+    };
+
+    const discount = {
+      actionKey: "test",
+      promotionId: "promotion1",
+      discountType: "item",
+      discountCalculationType: "fixed",
+      discountValue: 10,
+      discountedAmount: 2
+    };
+
+    item.discounts.push(discount);
+
+    mockContext.discountCalculationMethods = {
+      fixed: jest.fn().mockReturnValue(2)
+    };
+
+    applyItemDiscountToCart.recalculateCartItemSubtotal(mockContext, item);
+
+    expect(item.subtotal).toEqual({
+      amount: 10,
+      currencyCode: "USD",
+      discount: 2,
+      undiscountedAmount: 12
+    });
+  });
+
+  test("should recalculate the item subtotal with discountType is order", () => {
+    const item = {
+      _id: "item1",
+      price: {
+        amount: 12
+      },
+      quantity: 1,
+      subtotal: {
+        amount: 10,
+        currencyCode: "USD",
+        discount: 2,
+        undiscountedAmount: 12
+      },
+      discounts: []
+    };
+
+    const discount = {
+      actionKey: "test",
+      promotionId: "promotion1",
+      discountType: "order",
+      discountCalculationType: "fixed",
+      discountValue: 10,
+      discountedAmount: 5
+    };
+
+    item.discounts.push(discount);
+
+    applyItemDiscountToCart.recalculateCartItemSubtotal(mockContext, item);
+
+    expect(item.subtotal).toEqual({
+      amount: 7,
+      currencyCode: "USD",
+      discount: 5,
+      undiscountedAmount: 12
+    });
+  });
+});
