--- packages/api-plugin-promotions-discounts/src/enhancers/index.js
@@ -1,3 +1,0 @@
-import resetCartDiscountState from "./resetCartDiscountState.js";
-
-export default [resetCartDiscountState];
