+++ packages/api-plugin-promotions-discounts/src/index.js
@@ -1,17 +1,16 @@
 import { createRequire } from "module";
-import setDiscountsOnCart from "./utils/setDiscountsOnCart.js";
 import actions from "./actions/index.js";
 import methods from "./methods/index.js";
-import enhancers from "./enhancers/index.js";
-import addDiscountToOrderItem from "./utils/discountTypes/item/addDiscountToOrderItem.js";
-import getCartDiscountTotal from "./utils/discountTypes/order/getCartDiscountTotal.js";
-import getItemDiscountTotal from "./utils/discountTypes/item/getItemDiscountTotal.js";
-import getShippingDiscountTotal from "./utils/discountTypes/shipping/getShippingDiscountTotal.js";
-import getGroupDiscountTotal from "./utils/discountTypes/shipping/getGroupDisountTotal.js";
-import applyDiscountsToRates from "./utils/discountTypes/shipping/applyDiscountsToRates.js";
+import queries from "./queries/index.js";
+// import getCartDiscountTotal from "./utils/getCartDiscountTotal.js";
+// import getItemDiscountTotal from "./utils/getItemDiscountTotal.js";
+// import getShippingDiscountTotal from "./discountTypes/shipping/getShippingDiscountTotal.js";
+import getGroupDiscountTotal from "./discountTypes/shipping/getGroupDisountTotal.js";
+import applyDiscountsToRates from "./discountTypes/shipping/applyDiscountsToRates.js";
+import addDiscountToOrderItem from "./utils/addDiscountToOrderItem.js";
 import preStartup from "./preStartup.js";
-import recalculateDiscounts from "./xforms/recalculateDiscounts.js";
 import { discountCalculationMethods, registerDiscountCalculationMethod } from "./registration.js";
+import getTotalDiscountOnCart from "./utils/getTotalDiscountOnCart.js";
 
 const require = createRequire(import.meta.url);
 const pkg = require("../package.json");
@@ -30,30 +29,16 @@ export default async function register(app) {
       registerPluginHandler: [registerDiscountCalculationMethod],
       preStartup: [preStartup],
       mutateNewOrderItemBeforeCreate: [addDiscountToOrderItem],
-      calculateDiscountTotal: [getCartDiscountTotal, getItemDiscountTotal, getShippingDiscountTotal],
+      calculateDiscountTotal: [getTotalDiscountOnCart],
       getGroupDiscounts: [getGroupDiscountTotal],
       applyDiscountsToRates: [applyDiscountsToRates]
     },
-    cart: {
-      transforms: [
-        {
-          name: "setDiscountsOnCart",
-          fn: setDiscountsOnCart,
-          priority: 10
-        },
-        {
-          name: "recalculateDiscounts",
-          fn: recalculateDiscounts,
-          priority: 10
-        }
-      ]
-    },
+    queries,
     contextAdditions: {
       discountCalculationMethods
     },
     promotions: {
-      actions,
-      enhancers
+      actions
     },
     discountCalculationMethods: methods
   });
