+++ packages/api-plugin-promotions-discounts/src/preStartup.js
@@ -112,6 +112,10 @@ async function extendOrderSchemas(context) {
     }
   });
   OrderItem.extend({
+    "discount": {
+      type: Number,
+      optional: true
+    },
     "discounts": {
       type: Array,
       label: "Item Discounts",
