+++ packages/api-plugin-promotions-discounts/src/utils/getItemDiscountTotal.js
