--- packages/api-plugin-promotions-discounts/src/xforms/recalculateDiscounts.js
@@ -1,17 +1,0 @@
-import addDiscountToOrderItem from "../utils/discountTypes/item/addDiscountToOrderItem.js";
-
-/**
- * @summary Recalculates discounts on an order
- * @param {Object} context - The application context
- * @param {Object} cart - The cart to recalculate discounts on
- * @returns {void} undefined
- */
-export default function recalculateDiscounts(context, cart) {
-  // recalculate item discounts
-  for (const item of cart.items || []) {
-    addDiscountToOrderItem(context, { item, cartItem: item });
-  }
-
-  // TODO: Recalculate shipping discounts
-  // TODO: Recalculate order discounts
-}
