+++ packages/api-plugin-promotions-offers/src/simpleSchemas.js
@@ -1,14 +1,5 @@
 import SimpleSchema from "simpl-schema";
 
-const OfferTriggerFact = new SimpleSchema({
-  name: String,
-  handlerName: String,
-  fromFact: {
-    type: String,
-    optional: true
-  }
-});
-
 const Rules = new SimpleSchema({
   conditions: {
     type: Object,
@@ -17,23 +8,16 @@ const Rules = new SimpleSchema({
 });
 
 export const OfferTriggerParameters = new SimpleSchema({
-  "name": String,
-  "conditions": {
+  name: String,
+  conditions: {
     type: Object,
     blackbox: true
   },
-  "facts": {
-    type: Array,
-    optional: true
-  },
-  "facts.$": {
-    type: OfferTriggerFact
-  },
-  "inclusionRule": {
+  inclusionRule: {
     type: Rules,
     optional: true
   },
-  "exclusionRule": {
+  exclusionRule: {
     type: Rules,
     optional: true
   }
